pyproject.toml
[project]
name = "musicstreaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.43",
    "fastapi>=0.128.2",
    "faster-whisper>=1.2.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
